% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_breaks}
\alias{base_breaks}
\title{Pretty ticks for log scale in ggplot}
\usage{
base_breaks(n = 10)
}
\arguments{
\item{n}{Number of breaks}
}
\value{
a call to\code{\link[grDevices]{axisTicks}} to create pretty axis ticks.
}
\description{
Pretty ticks for log scale in ggplot
}
\examples{
\dontrun{
#Define values
n <- seq(1:10)
x <- 10*(1/2)^n
y <- gompertz(time = x, A = 2, mu = 0.6, lambda = 3)

df <- data.frame(x = x,
                 y = y)

ggplot(df, aes(x = x, y = y)) +
       geom_point() +
       scale_x_log10(breaks = base_breaks(n = 10)) +
       scale_y_log10(breaks = base_breaks(n = 6))
}

}
