% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{big_palette}
\alias{big_palette}
\title{A palette with 40 colors}
\format{
An object of class \code{character} of length 50.
}
\usage{
big_palette
}
\description{
This palette is used in the package for plotting.
}
\keyword{datasets}
