% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calibrate}
\alias{calibrate}
\title{Apply data transformation/calibration to measurement values in a dataframe}
\usage{
calibrate(df, equation)
}
\arguments{
\item{df}{An R dataframe object data in QurvE custom format.
The first three table rows contain
\enumerate{
\item sample description
\item replicate number (\emph{optional}: followed by a letter to indicate technical replicates)
\item concentration value (\emph{optional})
}}

\item{equation}{An equation in the form 'y = function(x)' (for example: 'y = x^2 * 0.3 - 0.5') to convert values.}
}
\value{
Dataframe \code{df} with values adjusted based on the provided equation.
}
\description{
Apply data transformation/calibration to measurement values in a dataframe
}
\examples{
\dontrun{
# Create random growth dataset
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = 'Test1')

# Convert dataset into QurvE custom format
df <- cbind(
  "time" = c("Time", NA, NA, rnd.data1$time[1,]),
  t(rnd.data1$data))

# Perform calibration
df.calib <- calibrate(df, "y = 2 * x")

# Inspect results
head(df)
head(df.calib)
}
}
