% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_RData}
\alias{export_RData}
\title{Export an R object as .RData file}
\usage{
export_RData(object, out.dir = tempdir(), out.nm = class(object))
}
\arguments{
\item{object}{An R object.}

\item{out.dir}{The path to the output directory. Default: the working directory}

\item{out.nm}{The output filename (with or without '.RData' ending). Default: the class of \code{object} followed by '.RData'.}
}
\value{
\code{NULL}
}
\description{
Export an R object as .RData file
}
\examples{
if(interactive()){
df <- data.frame('A' = seq(1:10), 'B' = rev(seq(1:10)))

export_RData(df)
}
}
