% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{fast.write.csv}
\alias{fast.write.csv}
\title{Write object in CSV file}
\usage{
fast.write.csv(dat, file, row.names = TRUE)
}
\arguments{
\item{dat}{An R data object (e.g., list, data frame)}

\item{file}{(Character) The name of the CSV file.}

\item{row.names}{(Logical) Add row names as column (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Write object in CSV file
}
