% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{fl.control}
\alias{fl.control}
\title{Title}
\usage{
fl.control(
  fit.opt = c("l", "s"),
  x_type = c("density", "time"),
  norm_fl = TRUE,
  t0 = 0,
  min.density = NA,
  log.x.lin = FALSE,
  log.x.spline = FALSE,
  log.y.lin = FALSE,
  log.y.spline = FALSE,
  lin.h = NULL,
  lin.R2 = 0.97,
  lin.RSD = 0.05,
  lin.dY = 0.05,
  dr.parameter = "max_slope.spline",
  dr.method = c("model", "spline"),
  dr.have.atleast = 5,
  smooth.dr = NULL,
  log.x.dr = FALSE,
  log.y.dr = FALSE,
  nboot.dr = 0,
  biphasic = FALSE,
  interactive = FALSE,
  nboot.fl = 0,
  smooth.fl = 0.75,
  growth.thresh = 1.5,
  suppress.messages = FALSE,
  neg.nan.act = FALSE,
  clean.bootstrap = TRUE
)
}
\arguments{
\item{dr.method}{For model, cite: https://doi.org/10.1038/s41467-020-20094-3}

\item{clean.bootstrap}{}
}
\description{
Title
}
