% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{fl.drFit}
\alias{fl.drFit}
\title{Title}
\usage{
fl.drFit(FitData, control = fl.control())
}
\arguments{
\item{control}{}
}
\description{
Title
}
\references{
Meyer, A.J., Segall-Shapiro, T.H., Glassey, E. et al. \emph{Escherichia coli “Marionette” strains with 12 highly optimized small-molecule sensors.} Nat Chem Biol 15, 196–204 (2019). \doi{10.1038/s41589-018-0168-3}
}
