% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{fl.report}
\alias{fl.report}
\title{Create a PDF and HTML report with results from a fluorescence analysis workflow}
\usage{
fl.report(flFitRes, report.dir = NULL, ec50, format = c("pdf", "html"), ...)
}
\arguments{
\item{flFitRes}{A \code{grofit} object created with \code{\link{fl.workflow}}.}

\item{report.dir}{(Character) The path or name of the folder in which the report files are created.  If \code{NULL}, the folder will be named with a combination of "Report.growth_" and the current date and time.}

\item{ec50}{(Logical) Display results of dose-response analysis (\code{TRUE}) or not (\code{FALSE}).}

\item{...}{Further arguments passed to create a report. Currently required:
\itemize{
\item \code{ec50}: \code{TRUE} or \code{FALSE}: Was a dose-response analysis performed in \code{\link{fl.workflow}}?
\item \code{mean.grp}: Define groups to combine into common plots in the report based on sample identifiers. Partial matches with sample/group names are accepted. Can be \code{"all"}, a vector of strings, or a list of string vectors. Note: The maximum number of sample groups (with unique condition/concentration indicators) is 50. If you have more than 50 groups, option \code{"all"} will produce the error \code{! Insufficient values in manual scale. [Number] needed but only 50 provided}.
\item \code{mean.conc}: Define concentrations to combine into common plots in the  report. Can be a numeric vector, or a list of numeric vectors.
\item \code{export}: Shall all plots generated in the report be exported as individual PDF and PNG files \code{TRUE} or not \code{FALSE}?
}}

\item{format(Character)}{Define the file format for the report, PDF (\code{'pdf'}) and/or HTML (\code{'html'}). Default: (\code{c('pdf', 'html')})}
}
\description{
\code{fl.report} requires a \code{flFitRes} object and creates a report in PDF and HTML format that summarizes all results obtained.
}
