% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{flBootSpline}
\alias{flBootSpline}
\title{flBootSpline: Function to generate a bootstrap}
\usage{
flBootSpline(
  time = NULL,
  density = NULL,
  fl_data,
  ID = "undefined",
  control = fl.control()
)
}
\arguments{
\item{ID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{fl.control} object created with \code{growth.control()}, defining relevant fitting options.}
}
\description{
flBootSpline: Function to generate a bootstrap
}
