% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{flFitSpline}
\alias{flFitSpline}
\title{Perform a bootstrap on a dataset containing fluorescence data followed by spline fits for each resample}
\usage{
flFitSpline(
  time = NULL,
  density = NULL,
  fl_data,
  ID = "undefined",
  control = fl.control(x_type = c("density", "time"), log.x.spline = FALSE,
    log.y.spline = FALSE, smooth.fl = 0.75, t0 = 0, min.density = NA)
)
}
\arguments{
\item{time}{Vector of the independent variable: time (if \code{x_type = 'time'} in \code{fl.control} object.}

\item{density}{Vector of the independent variable: density (if \code{x_type = 'density'} in \code{fl.control} object.}

\item{fl_data}{Vector of dependent variable: fluorescence.}

\item{ID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{fl.control} object created with \code{\link{growth.control}}, defining relevant fitting options.}
}
\value{
A \code{gcBootSpline} object containing a distribution of growth parameters and
a \code{gcFitSpline} object for each bootstrap sample. Use \code{\link{plot.gcBootSpline}}
to visualize all bootstrapping splines as well as the distribution of
physiological parameters.

\item{x.in}{Raw x values provided to the function as \code{time} or \code{density}.}
\item{fl.in}{Raw fluorescence data provided to the function as \code{fl_data}.}
\item{raw.x}{Filtered x values used for the spline fit.}
\item{raw.fl}{Filtered fluorescence values used for the spline fit.}
\item{ID}{(Character) Identifies the tested sample.}
\item{fit.x}{Fitted x values.}
\item{fit.fl}{Fitted fluorescence values.}
\item{parameters}{List of determined parameters.}
\itemize{
\item \code{A}: {Maximum fluorescence.}
\item \code{dY}: {Difference in maximum fluorescence and minimum fluorescence.}
\item \code{max_slope}: {Maximum slope of fluorescence-vs.-x data (i.e., maximum in first derivative of the spline).}
\item \code{x.max}: {Time at the maximum slope.}
\item \code{lambda}: {Lag time.}
\item \code{b.tangent}: {Intersection of the tangent at the maximum slope with the abscissa.}
\item \code{max_slope2}: {For biphasic course of fluorescence: Maximum slope of fluorescence-vs.-x data of the second phase.}
\item \code{lambda2}: {For biphasic course of fluorescence: Lag time determined for the second phase.}
\item \code{x.max2}: {For biphasic course of fluorescence: Time at the maximum slope of the second phase.}
\item \code{b.tangent2}: {For biphasic course of fluorescence: Intersection of the tangent at the maximum slope of the second phase with the abscissa.}
\item \code{integral}: {Area under the curve of the spline fit.}
}
\item{spline}{\code{smooth.spline} object generated by the \code{\link{smooth.spline}} function.}
\item{spline.deriv1}{list of time ('x') and density ('y') values describing the first derivative of the spline fit.}
\item{reliable}{(Logical) Indicates whether the performed fit is reliable (to be set manually).}
\item{fitFlag}{(Logical) Indicates whether a spline fit was successfully performed on the data.}
\item{fitFlag2}{(Logical) Indicates whether a second phase was identified.}
\item{control}{Object of class \code{fl.control} containing list of options passed to the function as \code{control}.}
}
\description{
\code{flBootSpline} resamples the fluorescence-'x' value pairs in a dataset with replacement and performs a spline fit for each bootstrap sample.
}
\seealso{
Other fluorescence fitting functions: 
\code{\link{flBootSpline}()},
\code{\link{flFit}()}
}
\concept{fluorescence fitting functions}
