% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tables.R
\name{get_sd_param}
\alias{get_sd_param}
\title{Calculate parameter standard deviations for several samples}
\usage{
get_sd_param(table = data.frame(), ndx.rep = list(), param1, param2 = NULL)
}
\arguments{
\item{table}{A dataframe with class \code{gcTable} or \code{flTable}, generated with \code{\link{growth.gcFit}} or \code{\link{flFit}}, respectively.}

\item{ndx.rep}{A list with each element being a vector of row indices for samples (replicates) for which to calculate standard deviations.}

\item{param1}{Name of the parameter for which to calculate standard deviations}

\item{param2}{(optional) Name of a second parameter for which to calculate standard deviations The results for the second parameter will be attached in parentheses after the first parameter.}
}
\value{
A vector of strings containing parameter standard deviations.
}
\description{
Internal function used within the \code{table_group} set of functions.
}
\examples{
\dontrun{
# Create random growth data set
rnd.data <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")


# Run growth curve analysis workflow
res <- growth.workflow(time = rnd.data$time,
                       data = rnd.data$data,
                       fit.opt = "s",
                       ec50 = FALSE,
                       export.res = FALSE,
                       parallelize = FALSE,
                       suppress.messages = TRUE)

conc <- unique(res$expdesign$concentration)
ndx.rep <- lapply(1:length(conc), function(x) grep(paste0("Test1.+", conc[x]), res$expdesign$label))

get_sd_param(res$gcFit$gcTable, ndx.rep, "mu.spline")
}

}
