% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{gompertz.exp}
\alias{gompertz.exp}
\title{Calculate the values of the modified Gompertz growth model for given time points and growth parameters.}
\usage{
gompertz.exp(time, A, mu, lambda, addpar)
}
\arguments{
\item{time}{A vector of time values}

\item{A}{Maximum density value}

\item{mu}{Growth rate}

\item{lambda}{Lag time}

\item{addpar}{Numeric vector of size two, addpar\link{1} corresponds to scaling parameter \eqn{\alpha} and addpar\link{2} corresponds to shifting parameter \eqn{t_{shift}}.}
}
\value{
A vector of density values
}
\description{
Calculate the values of the modified Gompertz growth model for given time points and growth parameters.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
density <- gompertz.exp(time, A = 3, mu=0.6, lambda = 3, addpar = c(0.1, 12))

plot(time, density)
}
