% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{grofit.param}
\alias{grofit.param}
\title{internal function called within \code{growth.gcFitModel()}.}
\usage{
grofit.param(time, data, gcID = "undefined", control)
}
\arguments{
\item{control}{}
}
\description{
internal function called within \code{growth.gcFitModel()}.
}
