% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{grow_exponential}
\alias{grow_exponential}
\title{Calculate the values of the exponential growth model for given time points and growth parameters.}
\usage{
grow_exponential(time, parms)
}
\arguments{
\item{time}{A vector of time values}

\item{parms}{A vector of length 2 with parameters: 1. start density value (y0), 2. exponential growth rate (mumax)}
}
\value{
A dataframe with time and calculated density columns
}
\description{
Calculate the values of the exponential growth model for given time points and growth parameters.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
data.exp <- grow_exponential(time, parms = c(0.05, 0.6))
plot(data.exp)
}
