% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{grow_linear}
\alias{grow_linear}
\title{Calculate the values of the linear growth model for given time points and growth parameters.}
\usage{
grow_linear(time, parms)
}
\arguments{
\item{time}{A vector of time values}

\item{parms}{A vector of length 2 with parameters: 1. start growth value (y0), 2. growth rate as slope of the linear model (mumax)}
}
\value{
A dataframe with time and calculated growth columns
}
\description{
Calculate the values of the linear growth model for given time points and growth parameters.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
data.lin <- grow_linear(time, parms = c(0.05, 0.6))
plot(data.lin)
}
