% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.drFit}
\alias{growth.drFit}
\title{Perform smooth spline fits on response vs. concentration data}
\usage{
growth.drFit(gcTable, control = growth.control())
}
\arguments{
\item{gcTable}{A dataframe containing the data for the dose-response curve estimation. Such table of class \code{gcTable} can be obtained by running \code{\link{growth.gcFit}}.}

\item{control}{A \code{grofit.control} object created with \code{\link{growth.control}}, defining relevant fitting options.}
}
\value{
An object of class \code{drFit}.
\item{raw.data}{Data that passed to the function as \code{gcTable}.}
\item{drTable}{Dataframe containing condition identifiers, fit options, and results of the dose-response analysis.}
\item{drBootSplines}{List of all \code{drBootSpline} objects generated by the call of \code{\link{growth.drBootSpline}} for each distinct experiment.}
\item{drFittedSplines}{List of all \code{drFitSpline} objects generated by the call of \code{\link{growth.drFitSpline}} for each distinct experiment.}
\item{control}{Object of class \code{grofit.control} containing list of options passed to the function as \code{control}.}
}
\description{
\code{growth.drFit} serves to determine dose-response curves on every condition in a
dataset. The response parameter can be chosen from every physiological parameter in a
\code{gcTable} table which is obtained via \code{\link{growth.gcFit}}. \code{\link{growth.drFit}}
calls the functions \code{\link{growth.drFitSpline}} and \code{\link{growth.drBootSpline}} to
generate a table with estimates for EC50 and respecting statistics.
}
\references{
Matthias Kahm, Guido Hasenbrink, Hella Lichtenberg-Frate, Jost Ludwig, Maik Kschischo (2010). \emph{grofit: Fitting Biological Growth Curves with R}. Journal of Statistical Software, 33(7), 1-21. DOI: 10.18637/jss.v033.i07
}
\seealso{
Other growth fitting functions: 
\code{\link{growth.gcBootSpline}()},
\code{\link{growth.gcFitLinear}()},
\code{\link{growth.gcFitModel}()},
\code{\link{growth.gcFitSpline}()},
\code{\link{growth.gcFit}()},
\code{\link{growth.workflow}()}
}
\concept{growth fitting functions}
