% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcBootSpline}
\alias{growth.gcBootSpline}
\title{gcBootSpline: Function to generate a bootstrap}
\usage{
growth.gcBootSpline(time, data, gcID = "undefined", control = growth.control())
}
\arguments{
\item{control}{}
}
\description{
gcBootSpline: Function to generate a bootstrap
}
