% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcFit}
\alias{growth.gcFit}
\title{Perform a growth curve analysis on all samples in a provided dataset}
\usage{
growth.gcFit(
  time,
  data,
  control = growth.control(),
  t0 = 0,
  lin.h = NULL,
  lin.R2 = 0.95,
  lin.RSD = 0.05,
  lin.dY = 0.05
)
}
\arguments{
\item{time}{(optional) A matrix containing time values for each sample.}

\item{data}{Either a \code{grodata} object created with \code{growth.read_data()}, a list containing a \code{'time'} matrix as well as \code{'density'} and, if appropriate, \code{'fluorescence1'} and \code{'fluorescence2'} dataframes, or a dataframe containing growth data (if a \code{time} matrix is provided as separate argument).}

\item{control}{A \code{grofit.control} object created with}

\item{t0}{(Numeric) Minimum time value considered for linear and spline fits.}

\item{lin.h}{(Numeric) Manually define the size of the sliding window used in \code{growth.gcFitLinear()}. If \code{NULL}, h is calculated for each samples based on the number of measurements in the growth phase of the plot.}

\item{lin.R2}{(Numeric) \ifelse{html}{\out{R<sup>2</sup>}}{\eqn{R^2}} threshold for \code{growth.gcFitLinear()}.}

\item{lin.RSD}{(Numeric) Relative standard deviation (RSD) threshold for \code{growth.gcFitLinear()}.}

\item{lin.dY}{(Numeric) Threshold for the minimum fraction of density increase a linear regression window should cover. Default: 0.05 (5\%).}
}
\value{
A \code{gcFit} object that contains all growth fitting results, compatible with various plotting functions of the QurvE package.
}
\description{
\code{growth.gcFit} performs all computational growth fitting operations based on the user input.
}
