% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcFitModel}
\alias{growth.gcFitModel}
\title{Fit different growth models to density data}
\usage{
growth.gcFitModel(time, data, gcID = "undefined", control = growth.control())
}
\arguments{
\item{control}{}
}
\description{
Fit different growth models to density data
}
