% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcFitSpline}
\alias{growth.gcFitSpline}
\title{Model a smooth spline fit for density data.}
\usage{
growth.gcFitSpline(
  time,
  data,
  gcID = "undefined",
  control = growth.control(),
  t0 = 0
)
}
\arguments{
\item{t0}{}
}
\description{
Model a smooth spline fit for density data.
}
