% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.parse_data}
\alias{growth.parse_data}
\title{Parse raw plate reader data and convert it to a format compatible with QurvE}
\usage{
growth.parse_data(
  file = NULL,
  map.file = NULL,
  software = "Gen5",
  convert.time = TRUE,
  sheet = 1,
  csvsep = ";",
  dec = ".",
  subtract.blank = T
)
}
\arguments{
\item{convert.time}{}
}
\description{
\code{growth.parse_data} takes a raw export file from a plate reader experiment, extracts relevant information and parses it into the format required to run \code{growth.workflow()}.
}
