% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.read_data}
\alias{growth.read_data}
\title{Read growth data in table format}
\usage{
growth.read_data(data, data.format = "col", csvsep = ";")
}
\arguments{
\item{data}{vector of dependent variable (concentration of organisms).}

\item{time}{vector of independent variable.}

\item{h}{width of the window (number of data).}

\item{quota}{part of window fits considered for the overall linear fit
(relative to max. growth rate)}
}
\value{
object with parameters of the fit. The lag time is currently estimated
as the intersection between the fit and the horizontal line with \eqn{y=y_0},
where \code{y0} is the first value of the dependent variable. The intersection
of the fit with the abscissa is indicated as \code{y0_lm} (lm for linear model).
These identifieres and their assumptions may change in future versions.
}
\description{
Reads a table file or R dataframe object containing growth data and extracts sample information, samp
}
\details{
The algorithm works as follows:
\enumerate{
\item Fit linear regressions to all subsets of \code{h} consecutive data
points (sliding window). If for example \eqn{h=5}, fit a linear regression to points
1 \dots 5, 2 \dots 6, 3 \dots 7 and so on. The method seeks the highest
rate of exponential growth, so the dependent variable is of course
log-transformed.
\item Find the subset with the highest slope \eqn{b_{max}} and
include also the data points of adjacent subsets that have a slope of
at least \eqn{quota \cdot b_{max}},
e.g. all data sets that have at least 95\\% of the maximum slope.
\item Fit a new linear model to the extended data window identified in step 2.
}
}
\references{
Hall, BG., Acar, H, Nandipati, A and Barlow, M (2014) Growth Rates Made Easy.
Mol. Biol. Evol. 31: 232-38, \doi{10.1093/molbev/mst187}
}
\seealso{
Other fitting functions: 
\code{\link{growth.gcFitLinear}()}
}
\concept{fitting functions}
