% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.report}
\alias{growth.report}
\title{Create a PDF and HTML report with results from a growth curve analysis workflow}
\usage{
growth.report(grofit, report.dir = NULL, ...)
}
\arguments{
\item{grofit}{A \code{grofit} object created with \code{growth.workflow()}.}

\item{report.dir}{(Character) The path or name of the folder in which the report files are created.  If \code{NULL}, the folder will be named with a combination of "Report.growth_" and the current date and time.}
}
\description{
\code{growth.report()} requires a \code{grofit} object and creates a report in PDF and HTML format that summarizes all results obtained.
}
