% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{help_modal}
\alias{help_modal}
\title{Adjusted modalDialog()  function to include classes}
\usage{
help_modal(
  ...,
  title = NULL,
  footer = NULL,
  size = c("m", "s", "l"),
  easyClose = TRUE,
  fade = TRUE,
  idcss = ""
)
}
\arguments{
\item{...}{UI elements for the body of the modal dialog box.}

\item{title}{An optional title for the dialog.}

\item{footer}{UI for footer. Use NULL for no footer.}

\item{size}{One of "s" for small, "m" (the default) for medium, or "l" for large.}

\item{easyClose}{If \code{TRUE}, the modal dialog can be dismissed by clicking outside the dialog box, or be pressing the Escape key. If \code{FALSE} (the default), the modal dialog can't be dismissed in those ways; instead it must be dismissed by clicking on a modalButton(), or from a call to removeModal() on the server.}

\item{fade}{If \code{FALSE}, the modal dialog will have no fade-in animation (it will simply appear rather than fade in to view).}

\item{idcss}{CSS class of the modal}
}
\value{
\code{div} HTML tag for a modal dialog
}
\description{
Adjusted modalDialog()  function to include classes
}
\examples{
\dontrun{
require(shiny)

shinyApp(
  ui = basicPage(
    tags$style(
      type = 'text/css',
      '.modal-dialog.test{ width: fit-content !important; }'
    ),
    actionButton("show", "Show modal dialog"),
    actionButton("shows", "Show modal dialog2")
  ),

  server = function(input, output) {

    # Render DT
    output$dt <- DT::renderDT(cbind(iris, iris))

    # Modal management
    observeEvent(input$show, {
      showModal(
        help_modal(easyClose = T,

                DT::DTOutput("dt"),

                footer = tagList(
                  modalButton("Cancel"),
                  actionButton("ok", "OK")
                ),
                idcss = "test"
        )
      )
    })

    observeEvent(input$shows, {
      showModal(
        help_modal(easyClose = T,

                DT::DTOutput("dt"),

                footer = tagList(
                  modalButton("Cancel"),
                  actionButton("ok", "OK")
                ),
                idcss = "tests"
        )
      )
    })
  }
)
}

}
\author{
Stackoverflow user \code{mfindinge}
}
