% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{initbaranyi}
\alias{initbaranyi}
\title{Generate initial values for parameter estimation with the Baranyi's growth model}
\usage{
initbaranyi(time, y, A, mu, lambda)
}
\arguments{
\item{time}{A vector of time values.}

\item{y}{A vector with growth values (e.g., density).}

\item{A}{Maximum of the curve. If a vector is provided only the first entry is used.}

\item{mu}{Maximum slope. If a vector is provided only the first entry is used.}

\item{lambda}{Lag time. If a vector is provided only the first entry is used.}
}
\value{
The parameters input as arguments in a list:
\item{A}{Maximum of the curve.}
\item{mu}{Maximum slope.}
\item{lambda}{Lag time.}
\item{addpar}{Minimum of the curve \eqn{y0}.}
}
\description{
\code{initlogistic} receives time points, growth data, values for \eqn{A}, \eqn{\mu} and \eqn{\lambda} and returns a list object which entries are used as initial values in the nonlinear fit procedure \code{\link{nls}} with the Baranyi's growth model.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
# Simulate baranyi growth curve
y <- baranyi(time, A = 3, mu=0.6, lambda = 2, addpar = 0.01)

init <- initbaranyi(time, y, A = 3, mu = 0.6, lambda = 2)

}
