% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initbiosensor}
\alias{initbiosensor}
\title{Generate initial values for parameter estimation with the biosensor response model}
\usage{
initbiosensor(x, y, n)
}
\arguments{
\item{x}{A vector of concentration values with at least four elements.}

\item{y}{A vector of response values with at least four elements.}

\item{n}{Cooperativity parameter}
}
\value{
A list:
\item{y.max}{Maximum response value.}
\item{K}{Sensitivity parameter.}
\item{n}{The initially defined cooperativity parameter.}
}
\description{
Generate initial values for parameter estimation with the biosensor response model
}
\examples{
\dontrun{
n <- seq(1:10)
conc <- (10*(1/2)^n)
response <- log10(y <- gompertz(conc, A = 2, mu = 0.6, lambda = 3))

init <- initbiosensor(x = conc,
                      y = response,
                      n= 0.5)
}

}
