% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{logistic}
\alias{logistic}
\title{Calculate the values of the logistic growth model for given time points and growth parameters.}
\usage{
logistic(time, A, mu, lambda, addpar = NULL)
}
\arguments{
\item{time}{A vector of time values}

\item{A}{Maximum growth value}

\item{mu}{Growth rate}

\item{lambda}{Lag time}

\item{addpar}{Additional parameters have no effect in this type of model. They belong to the standard model description in \code{QurvE} and are initialized as \code{addpar=NULL} in the function header.}
}
\value{
A vector of growth values
}
\description{
Calculate the values of the logistic growth model for given time points and growth parameters.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
# Simulate logistic growth curve
growth <- logistic(time, A = 3, mu=0.6, lambda = 3, addpar = 0.01)

plot(time, growth)
}
