% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_arg}
\alias{match_arg}
\title{match arg to choices}
\usage{
match_arg(arg, choices, multiple = FALSE, ignore_case = TRUE, trim_ws = FALSE)
}
\arguments{
\item{arg}{a character vector (of length one unless several.ok is TRUE) or
NULL}

\item{choices}{a character vector of candidate values}

\item{multiple}{logical specifying if arg should be allowed to have more than
one element. Defaults to FALSE}

\item{ignore_case}{logical indicating whether to ignore capitalization.
Defaults to TRUE}

\item{trim_ws}{logical indicating whether to trim surrounding white space.
Defaults to TRUE}
}
\value{
Value(s) matched via partial matching.
}
\description{
Wrapper around \link[base]{match.arg} that defaults to ignoring case and
trimming white space
}
\author{
Alban Sagouis
}
