% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_biolector_shiny}
\alias{parse_biolector_shiny}
\title{Extract relevant data from a raw data export file generated from the software of "Biolector" plate readers.}
\usage{
parse_biolector_shiny(input, density.nm)
}
\arguments{
\item{input}{A dataframe created by reading a table file with \code{\link{read_file}}}

\item{density.nm}{Name of read corresponding to growth rate}
}
\value{
a list of length two containing a density dataframe in the first element and \code{NA} in the second. The first column in the dataframe represents a time vector.
}
\description{
Extract relevant data from a raw data export file generated from the software of "Biolector" plate readers.
}
\examples{
\dontrun{
input <- read_file(filename = system.file("biolector", package = "QurvE"), csvsep = "," )
parsed <- parse_biolector_shiny(input, "[1] Biomass Gain=3")
}
}
