% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_chibio_shiny}
\alias{parse_chibio_shiny}
\title{Extract relevant data from a raw data export file generated from the software of "Chi.Bio" bioreactors.}
\usage{
parse_chibio_shiny(input, density.nm, fl.nm, fl2.nm)
}
\arguments{
\item{input}{A dataframe created by reading a table file with \code{\link{read_file}}}

\item{density.nm}{Name of read corresponding to growth rate}

\item{fl.nm}{Name of read corresponding to fluorescence data}

\item{fl2.nm}{Name of read corresponding to fluorescence 2 data}
}
\value{
a list of length two containing density and/or fluorescence dataframes in the first and second element, respectively. The first column in these dataframes represents a time vector.
}
\description{
Extract relevant data from a raw data export file generated from the software of "Chi.Bio" bioreactors.
}
\examples{
\dontrun{
input <- read_file(filename = system.file("ChiBio.csv", package = "QurvE"), csvsep = "," )
parsed <- parse_chibio(input, "od_measured", "FP1_emit1")
}
}
