% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{parse_data}
\alias{parse_data}
\title{Parse raw plate reader data and convert it to a format compatible with QurvE}
\usage{
parse_data(
  data.file = NULL,
  map.file = NULL,
  software = "Gen5",
  convert.time = TRUE,
  data.sheet = 1,
  map.sheet = 1,
  csvsep.data = ";",
  dec.data = ".",
  csvsep.map = ";",
  dec.map = ".",
  subtract.blank = T
)
}
\arguments{
\item{data.file}{(Character) A table file with extension '.xlsx', '.xls', '.csv', '.tsv', or '.txt' containing plate reader data.}

\item{map.file}{(Character) A table file in column format with 'well', 'ID', 'replicate', and 'concentration' in the first row. Used to assign sample information to wells in a plate.}

\item{software}{(Character) The name of the software used to export the plate reader data.}

\item{convert.time}{(Logical) Shall the time be converted from, e.g., 00:15:00 in Excel into hours (\code{TRUE}) or not (\code{FALSE})?}

\item{data.sheet}{(Numeric or Character) Number or name of a sheet in XLS or XLSX files containing experimental data (\emph{optional}).}

\item{map.sheet}{(Numeric or Character) Number or name of a sheet in XLS or XLSX files containing experimental data (\emph{optional}).}

\item{csvsep.data}{(Character) separator used in CSV data file (ignored for other file types).  Default: \code{";"}}

\item{dec.data}{(Character) decimal separator used in CSV, TSV or TXT data file.}

\item{csvsep.map}{(Character) separator used in CSV mapping file (ignored for other file types).  Default: \code{";"}}

\item{dec.map}{(Character) decimal separator used in CSV, TSV or TXT mapping file.}

\item{subtract.blank}{(Logical) Shall blank values be subtracted from values within the same experiment (\link{TRUE}, the default) or not (\link{FALSE}).}
}
\value{
A \code{grodata} object suitable to run \code{\link{growth.workflow}}. See \code{\link{read_data}} for object structure.
}
\description{
\code{parse_data} takes a raw export file from a plate reader experiment, extracts relevant information and parses it into the format required to run \code{\link{growth.workflow}}.
}
