% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_Gen5Gen6_shiny}
\alias{parse_Gen5Gen6_shiny}
\title{Extract relevant data from a raw data export file generated with the "Gen5" or "Gen6" software.}
\usage{
parse_Gen5Gen6_shiny(data, density.nm, fl.nm)
}
\arguments{
\item{data}{A dataframe created by reading a table file with \code{\link{read_file}}}

\item{density.nm}{Name of read corresponding to growth rate}

\item{fl.nm}{Name of read corresponding to fluorescence data}
}
\value{
a list of length two containing density and/or fluorescence dataframes in the first and second element, respectively. The first column in these dataframes represents a time vector, the remainder the measurements.
}
\description{
Extract relevant data from a raw data export file generated with the "Gen5" or "Gen6" software.
}
\examples{
\dontrun{
input <- read_file(filename = system.file("fluorescence_test_Gen5.xlsx", package = "QurvE") )
parsed <- parse_Gen5Gen6_shiny(input, "Read 3:630", "GFP:485,528")
}
}
