% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsers.R
\name{parse_growthprofiler}
\alias{parse_growthprofiler}
\title{Extract relevant data from a raw data export file generated from the software of the "Growth Profiler".}
\usage{
parse_growthprofiler(input)
}
\arguments{
\item{input}{A dataframe created by reading a table file with \code{\link{read_file}}}
}
\value{
a list of length two containing a density dataframe in the first element and \code{NA} in the second. The first column in the dataframe represents a time vector.
}
\description{
Extract relevant data from a raw data export file generated from the software of the "Growth Profiler".
}
\examples{
\dontrun{
input <- read_file(filename = system.file("GrowthProfiler.csv", package = "QurvE"), csvsep = "," )
parsed <- parse_growthprofiler(input)
}
}
