% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_properties_biolector}
\alias{parse_properties_biolector}
\title{Extract names of biomass reads from experimental data created with a Biolector plate reader}
\usage{
parse_properties_biolector(file, csvsep = ";", dec = ".", sheet = 1)
}
\arguments{
\item{file}{path to file with experiment data}

\item{csvsep}{CSV separator as string.}

\item{dec}{Decimal separator as string (for CSV, TXT, and TSV files)}

\item{sheet}{Sheet number or name (as string) for XLS and XLSX files.}
}
\value{
A vector of read name strings
}
\description{
Extract names of biomass reads from experimental data created with a Biolector plate reader
}
\examples{
\dontrun{
parse_properties_biolector(file = system.file("biolector.csv", package = "QurvE"), csvsep = ";")
}
}
