% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_properties_Gen5Gen6}
\alias{parse_properties_Gen5Gen6}
\title{Extract names of reads from experimental data created with the "Gen5" or "Gen6" software}
\usage{
parse_properties_Gen5Gen6(file, csvsep = ";", dec = ".", sheet = 1)
}
\arguments{
\item{file}{path to file with experiment data}

\item{csvsep}{CSV separator as string.}

\item{dec}{Decimal separator as string (for CSV, TXT, and TSV files)}

\item{sheet}{Sheet number or name (as string) for XLS and XLSX files.}
}
\value{
A vector of read name strings
}
\description{
Extract names of reads from experimental data created with the "Gen5" or "Gen6" software
}
\examples{
\dontrun{
parse_properties_Gen5Gen6(file = system.file("fluorescence_test_Gen5.xlsx", package = "QurvE"))
}
}
