% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_tecan_shiny}
\alias{parse_tecan_shiny}
\title{Extract relevant data from a raw data export file generated from the software of "Tecan" plate readers.}
\usage{
parse_tecan_shiny(input, growth.nm, fl.nm, fl2.nm)
}
\arguments{
\item{input}{A dataframe created by reading a table file with \code{\link{read_file}}}

\item{growth.nm}{Name of read corresponding to growth rate}

\item{fl.nm, fl2.nm}{Name of read corresponding to fluorescence and fluorescence2 data}
}
\value{
a list of length two containing growth and/or fluorescence dataframes in the first and second element, respectively. The first column in these dataframes represents a time vector.
}
\description{
Extract relevant data from a raw data export file generated from the software of "Tecan" plate readers.
}
\examples{
\dontrun{
input <- read_file(filename = system.file("Tecan.csv", package = "QurvE"), csvsep = "," )
parsed <- parse_tecan(input, "Label1_Copy1", "sfGFP", "mRFP1")
}
}
