% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{parse_victorx3_shiny}
\alias{parse_victorx3_shiny}
\title{Extract relevant data from a raw data export file generated from the software of Perkin Elmer's "Victor X3" plate readers.}
\usage{
parse_victorx3_shiny(input, growth.nm, fl.nm, fl2.nm)
}
\arguments{
\item{input}{A dataframe created by reading a table file with \code{\link{read_file}}}

\item{growth.nm}{Name of read corresponding to growth rate}

\item{fl.nm, fl2.nm}{Name of read corresponding to fluorescence and fluorescence2 data}
}
\value{
a list of length two containing growth and/or fluorescence dataframes in the first and second element, respectively. The first column in these dataframes represents a time vector.
}
\description{
Extract relevant data from a raw data export file generated from the software of Perkin Elmer's "Victor X3" plate readers.
}
\examples{
\dontrun{
input <- read_file(filename = system.file("victorx3_output.txt", package = "QurvE") )
parsed <- parse_victorx3_shiny(input, "Absorbance @ 600 (A)", "GFP (Counts)")
}
}
