% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.drFit}
\alias{plot.drFit}
\title{Generic plot function for \code{drFit} objects.}
\usage{
\method{plot}{drFit}(
  drFit,
  combine = TRUE,
  pch = 16,
  cex = 2,
  basesize = 15,
  colors = NULL,
  lwd = 0.7,
  ec50line = TRUE,
  plot = TRUE,
  export = FALSE,
  height = NULL,
  width = NULL,
  out.dir = NULL,
  out.nm = NULL
)
}
\arguments{
\item{drFit}{object of class \code{drFit}, created with \code{growth.drFit()}.}

\item{combine}{(Logical) Combine the dose-response analysis results of all conditions into a single plot (\code{TRUE}) or not (\code{FALSE}).}

\item{pch}{(Numeric) Shape of the raw data symbols.}

\item{basesize}{(Numeric) Base font size.}

\item{colors}{(Numeric or character) Define colors for different conditions.}

\item{lwd}{(Numeric) Line width of the individual splines.}

\item{ec50line}{(Logical) Show pointed horizontal and vertical lines at the EC50 values (\code{TRUE}) or not (\code{FALSE}).}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}).}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{out.nm}{(Character) The name of the PDF and PNG files if \code{export = TRUE}. If \code{NULL}, a name will be automatically generated including the chosen parameter.}
}
\description{
code{plot.drFit} calls code{plot.drFitSpline} for each group used in a dose-response analysis
}
