% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R
\name{plot.drFitModel}
\alias{plot.drFitModel}
\title{Title}
\usage{
\method{plot}{drFitModel}(
  drFittedModel,
  ec50line = TRUE,
  log = c("x", "y"),
  pch = 1,
  colSpline = 1,
  colData = 1,
  cex = 1,
  lwd = 2,
  plot = TRUE,
  export = FALSE,
  height = 7,
  width = 9,
  out.dir = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
