% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R
\name{plot.dual}
\alias{plot.dual}
\title{Title}
\usage{
\method{plot}{dual}(
  object,
  fluorescence = c("fl1", "fl2", "norm.fl1", "norm.fl2"),
  names = NULL,
  conc = NULL,
  mean = TRUE,
  log.y.density = F,
  log.y.fl = F,
  deriv = F,
  n.ybreaks = 6,
  colors = NULL,
  basesize = 20,
  y.lim.density = NULL,
  y.lim.fl = NULL,
  x.lim = NULL,
  y.title = NULL,
  x.title = NULL,
  y.title.density = NULL,
  y.title.fl = NULL,
  lwd = 1.1,
  plot = TRUE,
  export = FALSE,
  height = NULL,
  width = NULL,
  out.dir = NULL,
  out.nm = NULL
)
}
\arguments{
\item{out.nm}{(Character) The name of the PDF and PNG files if \code{export = TRUE}. If \code{NULL}, a name will be automatically generated including the chosen parameter.}
}
\description{
Title
}
