% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R
\name{plot.flFitLinear}
\alias{plot.flFitLinear}
\title{Title}
\usage{
\method{plot}{flFitLinear}(
  flFittedLinear,
  log = "",
  which = c("fit", "diagnostics"),
  plot = TRUE,
  export = FALSE,
  height = ifelse(which == "fit", 7, 5),
  width = ifelse(which == "fit", 9, 9),
  out.dir = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
