% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R
\name{plot.flFitRes}
\alias{plot.flFitRes}
\title{Title}
\usage{
\method{plot}{flFitRes}(
  object,
  data.type = c("spline1", "spline2", "raw1", "raw2", "norm.fl1", "norm.fl2"),
  names = NULL,
  conc = NULL,
  mean = TRUE,
  log.y = F,
  deriv = F,
  n.ybreaks = 6,
  colors = NULL,
  basesize = 20,
  y.lim = NULL,
  x.lim = NULL,
  y.title = NULL,
  x.title = NULL,
  y.lim.deriv = NULL,
  y.title.deriv = NULL,
  lwd = 1.1,
  plot = TRUE,
  export = FALSE,
  height = NULL,
  width = NULL,
  out.dir = NULL
)
}
\arguments{
\item{out.dir}{}
}
\description{
Title
}
