% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R
\name{plot.flFitSpline}
\alias{plot.flFitSpline}
\title{Generic plot function for \code{flFitSpline} objects.}
\usage{
\method{plot}{flFitSpline}(
  flFitSpline,
  add = FALSE,
  raw = TRUE,
  slope = TRUE,
  deriv = T,
  spline = T,
  log.y = F,
  pch = 1,
  colData = 1,
  colSpline = "dodgerblue3",
  cex = 1,
  lwd = 0.7,
  y.lim = NULL,
  x.lim = NULL,
  y.lim.deriv = NULL,
  plot = TRUE,
  export = FALSE,
  width = 8,
  height = ifelse(deriv == TRUE, 8, 6),
  out.dir = NULL,
  ...
)
}
\arguments{
\item{flFitSpline}{object of class \code{flFitSpline}, created with \code{\link{flFitSpline}}.}

\item{add}{(Logical) Shall the fitted spline be added to an existing plot? \code{TRUE} is used internally by \code{\link{plot.flBootSpline}}.}

\item{raw}{(Logical) Display raw density as circles (\code{TRUE}) or not (\code{FALSE}).}

\item{slope}{(Logical) Show the slope at the maximum growth rate (\code{TRUE}) or not (\code{FALSE}).}

\item{deriv}{(Logical) Show the derivative (i.e., slope) over time in a secondary plot (\code{TRUE}) or not (\code{FALSE}).}

\item{spline}{(Logical) Only for \code{add = TRUE}: add the current spline to the existing plot (\code{FALSE}).}

\item{log.y}{(Logical) Log-transform the y-axis (\code{TRUE}) or not (\code{FALSE}).}

\item{pch}{(Numeric) Size of the raw data circles.}

\item{colData}{(Numeric or character) Contour color of the raw data circles.}

\item{colSpline}{(Numeric or character) Spline line colour.}

\item{lwd}{(Numeric) Spline line width.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}). If \code{FALSE}, a ggplot object is returned.}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{...}{}

\item{basesize}{(Numeric) Base font size.}
}
\description{
code{plot.flFitSpline} generates the spline fit plot for a single sample.
}
