% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.gcBootSpline}
\alias{plot.gcBootSpline}
\title{Generic plot function for \code{gcBootSpline} objects.}
\usage{
\method{plot}{gcBootSpline}(
  gcBootSpline,
  pch = 1,
  colData = 1,
  deriv = TRUE,
  colSpline = ggplot2::alpha("dodgerblue3", 0.2),
  cex = 1,
  plot = TRUE,
  export = FALSE,
  height = 7,
  width = 9,
  out.dir = NULL,
  ...
)
}
\arguments{
\item{gcBootSpline}{object of class \code{gcBootSpline}, created with \code{\link{growth.gcBootSpline}}.}

\item{pch}{(Numeric) Size of the raw data circles.}

\item{colData}{(Numeric or character) Contour color of the raw data circles.}

\item{deriv}{(Logical) Show the derivatives (i.e., slope) over time in a secondary plot (\code{TRUE}) or not (\code{FALSE}).}

\item{colSpline}{(Numeric or character) Spline line colour.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}).}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{...}{}
}
\description{
Generic plot function for \code{gcBootSpline} objects.
}
