% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.grofit}
\alias{plot.grofit}
\title{Combine different groups of samples into a single plot}
\usage{
\method{plot}{grofit}(
  grofit,
  names = NULL,
  conc = NULL,
  mean = TRUE,
  log.y = T,
  deriv = T,
  n.ybreaks = 6,
  colors = NULL,
  basesize = 20,
  lwd = 1.1,
  plot = TRUE,
  export = FALSE,
  height = ifelse(deriv == T, 9, 6),
  width = 10 + 3 * ifelse(mean == TRUE, length(conditions_unique), length(nm))/15,
  out.dir = NULL
)
}
\arguments{
\item{grofit}{A \code{grofit} object created with \code{growth.workflow()} containing spline fits.}

\item{names}{(String or string vector) Define groups to combine into a single plot. Partial matches with sample/group names are accepted. If \code{NULL}, all samples are considered.}

\item{conc}{(Numeric or numeric vector) Define concentrations to combine into a single plot. If \code{NULL}, all concentrations are considered.}

\item{mean}{(Logical) Display the mean and standard deviation of groups with replicates (\code{TRUE}) or plot each sample individually (\code{FALSE})?}

\item{log.y}{(Logical) Log-transform the y-axis of the plot (\code{TRUE}) or not (\code{FALSE})?}

\item{deriv}{(Logical) Show derivatives over time in a separate panel below the plot (\code{TRUE}) or not (\code{FALSE})?}

\item{n.ybreaks}{(Numeric) Number of breaks on the y-axis. The breaks are generated using \code{scales::pretty_breaks}. Thus, the final number of breaks can deviate from the user input.}

\item{colors}{(String vector) Define a color palette used to draw the plots. If \code{NULL}, default palettes are chosen based on the number of groups/samples within the plot. Note: The number of provided colors should at least match the number of groups/samples.}

\item{basesize}{(Numeric) Base font size.}

\item{lwd}{(Numeric) Line width of the individual plots.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}). If \code{FALSE}, a ggplot object is returned.}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}
}
\description{
\code{plot.grofit} extracts the spline fits of a subset of samples in a \code{grofit} object calculates averages and standard deviations of conditions with replicates and combines them into a single plot.
}
