% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.parameter}
\alias{plot.parameter}
\title{Compare growth parameters between samples or conditions}
\usage{
\method{plot}{parameter}(
  object,
  param = c("mu.linfit", "lambda.linfit", "dY.linfit", "A.linfit", "mu.model",
    "lambda.model", "A.model", "mu.spline", "lambda.spline", "A.spline", "dY.spline",
    "integral.spline", "mu.bt", "lambda.bt", "A.bt", "integral.bt"),
  names = NULL,
  conc = NULL,
  basesize = 12,
  plot = T,
  export = F,
  height = 7,
  width = NULL,
  out.dir = NULL
)
}
\arguments{
\item{object}{A \code{grofit}, \code{gcFit}, or \code{gcTable} object obtained with \code{growth.workflow()} or \code{growth.gcFit}.}

\item{param}{(Character) The parameter used to compare different sample groups. Any name of a column containing numeric values in \code{gcTable} (which is stored within \code{grofit} or \code{gcFit} objects) can be used as input. Useful options are:
'mu.linfit', 'lambda.linfit', 'dY.linfit', 'A.linfit',
'mu.model', 'lambda.model', 'A.model',
'mu.spline', 'lambda.spline', 'A.spline', 'dY.spline', 'integral.spline',
'mu.bt', 'lambda.bt', 'A.bt', 'integral.bt'}

\item{names}{(String or string vector) Define groups to combine into a single plot. Partial matches with sample/group names are accepted. If \code{NULL}, all samples are considered. Note: Ensure to use unique substrings to extract groups of interest. If the name of one condition is included in its entirety within the name of other conditions, it cannot be extracted individually.}

\item{conc}{(Numeric or numeric vector) Define concentrations to combine into a single plot. If \code{NULL}, all concentrations are considered. Note: Ensure to use unique concentration values to extract groups of interest. If the concentration value of one condition is included in its entirety within the name of other conditions (e.g., the dataset contains '1', '10', and '100', \code{code = 10} will select both '10 and '100'), it cannot be extracted individually.}

\item{basesize}{(Numeric) Base font size.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}). If \code{FALSE}, a ggplot object is returned.}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}
}
\description{
\code{plot.parameter} gathers physiological parameters from the results of a growth fit analysis and compares a chosen parameter between each sample or condition in a column plot.
}
