% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app_functions.R
\name{run_app}
\alias{run_app}
\title{Run Shiny QurvE App}
\usage{
run_app()
}
\value{
Launches a browser with the shiny app
}
\description{
Run Shiny QurvE App
}
\examples{
if(interactive()){
# Run the app
run_app()
}
}
