% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{single_hue_palettes}
\alias{single_hue_palettes}
\title{Single hue palettes for ggplot2}
\format{
An object of class \code{list} of length 11.
}
\usage{
single_hue_palettes
}
\description{
These palettes are used in the package for plotting.
}
\keyword{datasets}
