% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.drBootSpline}
\alias{summary.drBootSpline}
\title{Generic summary function for drBootSpline objects}
\usage{
\method{summary}{drBootSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{drBootSpline}}

\item{...}{}
}
\value{
A dataframe with statistical parameters extracted from the dose-response bootstrapping analysis.
}
\description{
Generic summary function for drBootSpline objects
}
