% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.drFit}
\alias{summary.drFit}
\title{Generic summary function for drFit objects}
\usage{
\method{summary}{drFit}(object, ...)
}
\arguments{
\item{object}{object of class \code{drFit}}

\item{...}{}
}
\value{
A dataframe with parameters for all samples extracted from the dose-response analysis.
}
\description{
Generic summary function for drFit objects
}
