% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R
\name{summary.drFitFLModel}
\alias{summary.drFitFLModel}
\title{Generic summary function for drFitFLModel objects}
\usage{
\method{summary}{drFitFLModel}(object, ...)
}
\arguments{
\item{object}{object of class \code{drFitModel}}

\item{...}{}
}
\value{
A dataframe with biosensor response parameters.
}
\description{
Generic summary function for drFitFLModel objects
}
