% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R, R/growth_summaries.R
\name{summary.drFitModel}
\alias{summary.drFitModel}
\title{Generic summary function for drFitModel objects}
\usage{
\method{summary}{drFitModel}(object, ...)

\method{summary}{drFitModel}(object, ...)
}
\arguments{
\item{object}{object of class \code{drFitModel}}

\item{...}{}
}
\value{
A dataframe with biosensor response parameters.

A dataframe with parameters extracted from the dose-response analysis of a single sample.
}
\description{
Generic summary function for drFitModel objects

Generic summary function for drFitModel objects
}
