% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.drFitSpline}
\alias{summary.drFitSpline}
\title{Generic summary function for drFitSpline objects}
\usage{
\method{summary}{drFitSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{drFitSpline}}

\item{...}{}
}
\value{
A dataframe with parameters extracted from the dose-response analysis of a single sample.
}
\description{
Generic summary function for drFitSpline objects
}
