% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R
\name{summary.flFit}
\alias{summary.flFit}
\title{Generic summary function for flFit objects}
\usage{
\method{summary}{flFit}(object, ...)
}
\arguments{
\item{object}{object of class \code{flFit}}

\item{...}{Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.}
}
\value{
A dataframe with parameters extracted from all fits of a workflow.
}
\description{
Generic summary function for flFit objects
}
