% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R
\name{summary.flFitLinear}
\alias{summary.flFitLinear}
\title{Generic summary function for flFitLinear objects}
\usage{
\method{summary}{flFitLinear}(object, ...)
}
\arguments{
\item{object}{object of class \code{flFitLinear}}

\item{...}{}
}
\value{
A dataframe with parameters extracted from the linear fit.
}
\description{
Generic summary function for flFitLinear objects
}
