% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.gcBootSpline}
\alias{summary.gcBootSpline}
\title{Generic summary function for gcBootSpline objects}
\usage{
\method{summary}{gcBootSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{gcBootSpline}}

\item{...}{Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.}
}
\value{
A dataframe with statistical parameters extracted from the spline fit bootstrapping computation.
}
\description{
Generic summary function for gcBootSpline objects
}
