% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.gcFit}
\alias{summary.gcFit}
\title{Generic summary function for gcFit objects}
\usage{
\method{summary}{gcFit}(object, ...)
}
\arguments{
\item{object}{object of class \code{gcFit}}

\item{...}{}
}
\value{
A dataframe with parameters extracted from all fits of a workflow.
}
\description{
Generic summary function for gcFit objects
}
