% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.gcFitLinear}
\alias{summary.gcFitLinear}
\title{Generic summary function for gcFitLinear objects}
\usage{
\method{summary}{gcFitLinear}(object, ...)
}
\arguments{
\item{object}{object of class \code{gcFitLinear}}

\item{...}{Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.}
}
\value{
A dataframe with parameters extracted from the linear fit.
}
\description{
Generic summary function for gcFitLinear objects
}
