% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.gcFitModel}
\alias{summary.gcFitModel}
\title{Generic summary function for gcFitModel objects}
\usage{
\method{summary}{gcFitModel}(object, ...)
}
\arguments{
\item{object}{object of class \code{gcFitModel}}

\item{...}{}
}
\value{
A dataframe with parameters extracted from the growth model fit.
}
\description{
Generic summary function for gcFitModel objects
}
