% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_summaries.R
\name{summary.gcFitSpline}
\alias{summary.gcFitSpline}
\title{Generic summary function for gcFitSpline objects}
\usage{
\method{summary}{gcFitSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{gcFitSpline}}

\item{...}{}
}
\value{
A dataframe with parameters extracted from the nonparametric fit.
}
\description{
Generic summary function for gcFitSpline objects
}
