% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fits.R
\name{theil_sen_regression}
\alias{theil_sen_regression}
\title{Compute Theil Sen Regression}
\usage{
theil_sen_regression(formula, data = NULL)
}
\arguments{
\item{formula}{A formula in the form \code{y ~ x}.}

\item{data}{optional: A dataframe in which numeric values for \code{x} and \code{y} are stored}
}
\value{
A model object of class \code{lm}.
}
\description{
From the 'RobustLinearReg' package
}
\examples{
\dontrun{
# Create random growth dataset
rnd.dataset <- rdm.data(d = 35, mu = 0.8, A = 5, label = "Test1")

# Extract time and growth data for single sample
time <- rnd.dataset$time[1,]
data <- as.numeric(rnd.dataset$data[1,-(1:3)]) # Remove identifier columns
data.log <- log(data/data[1])

# Perform theil-sen regression
regression <- theil_sen_regression(data.log ~ time)
summary(regression)

df <- data.frame("x" = time, "y" = data.log)
regression <- theil_sen_regression(y ~ x, df)

summary(regression)

}
}
\references{
Henri Theil and Pranab K. Sen, 1950 and 1968 respectively
}
\author{
Santiago I. Hurtado
}
