% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xgx_minor_breaks_log10}
\alias{xgx_minor_breaks_log10}
\title{Sets the default minor_breaks for log10 scales (from R package 'xgxr')}
\usage{
xgx_minor_breaks_log10(data_range)
}
\arguments{
\item{data_range}{range of the data}
}
\value{
numeric vector of breaks
}
\description{
\code{xgx_minor_breaks_log10} sets nice minor_breaks for log10 scale.
}
