% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{baranyi}
\alias{baranyi}
\title{Calculate the values of the Baranyi growth model for given time points and growth parameters.}
\usage{
baranyi(time, A, mu, lambda, addpar)
}
\arguments{
\item{time}{A vector of time values}

\item{A}{Maximum density value}

\item{mu}{Growth rate}

\item{lambda}{Lag time}

\item{addpar}{additional parameters (here: minimum density)}
}
\value{
A vector of density values
}
\description{
Calculate the values of the Baranyi growth model for given time points and growth parameters.
}
\examples{
time <- seq(from = 0, to = 24, by = 0.25)
density <- baranyi(time, A = 3, mu=0.6, lambda = 3, addpar = 0.01)

plot(time, density)
}
