% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{colFmt}
\alias{colFmt}
\title{Format font color for Markdown reports}
\usage{
colFmt(x, color)
}
\arguments{
\item{x}{A character string. The text to be colored.}

\item{color}{(Character) A color.}
}
\value{
A LaTeX- or HTML-formatted string to assign a color to text based on the output format.
}
\description{
\code{colFmt} formats the input depending on PDF or HTML output to give colored text in reports.
}
\examples{
\dontrun{
# the example below is meaningless unless you put it in a markdown document
colFmt("Test_text", "Red")
}
}
