% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{colFmt}
\alias{colFmt}
\title{Format font color for Markdown reports}
\usage{
colFmt(x, color)
}
\arguments{
\item{x}{A character string. The text to be colored.}

\item{color}{(Character) A color.}
}
\description{
\code{colFmt} formats the input depending on PDF or HTML output to give colored text in reports.
}
