% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_computation.R
\name{fl.drFit}
\alias{fl.drFit}
\title{Fit a biosensor model (Meyer et al., 2019) to response vs. concentration data}
\usage{
fl.drFit(FitData, control = fl.control())
}
\arguments{
\item{FitData}{A dataframe containing the data for the dose-response model estimation. Such table of class \code{flTable} can be obtained by running \code{\link{flFit}} with \code{dr.method = "model"} as argument in the \code{fl.control} object.}

\item{control}{A \code{fl.control} object created with \code{\link{fl.control}}, defining relevant fitting options.}
}
\value{
An object of class \code{drFit}.
\item{raw.data}{Data that passed to the function as \code{flTable}.}
\item{drTable}{Dataframe containing condition identifiers, fit options, and results of the dose-response analysis.}
\item{drFittedModels}{List of all \code{drFitModel} objects generated by the call of \code{\link{fl.drFitModel}} for each distinct experiment.}
\item{control}{Object of class \code{fl.control} created with the call of \code{\link{fl.control}}.}
}
\description{
Fit a biosensor model (Meyer et al., 2019) to response vs. concentration data
}
\references{
Meyer, A.J., Segall-Shapiro, T.H., Glassey, E. et al. \emph{Escherichia coli “Marionette” strains with 12 highly optimized small-molecule sensors.} Nat Chem Biol 15, 196–204 (2019). \doi{10.1038/s41589-018-0168-3}
}
