% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_workflows.R
\name{flFit}
\alias{flFit}
\title{Perform a fluorescence curve analysis on all samples in the provided dataset.}
\usage{
flFit(fl_data, time = NULL, density = NULL, control = fl.control(), ...)
}
\arguments{
\item{fl_data}{Either... \enumerate{ \item a \code{grodata} object created with \code{\link{read_data}} or \code{\link{parse_data}},
\item a list containing a \code{'time'} matrix (for x_type == "time") or \code{'density'} dataframe (for x_type == "density") and a \code{'fluorescence'} dataframes,
or \item a dataframe containing (normalized) fluorescence values (if a \code{time} matrix or \code{density} dataframe is provided as separate argument).}}

\item{time}{(optional) A matrix containing time values for each sample.}

\item{density}{(optional) A dataframe containing density values for each sample and sample identifiers in the first three columns.}

\item{control}{A \code{fl.control} object created with \code{\link{fl.control}},
defining relevant fitting options.}

\item{...}{Further arguments passed to the shiny app.}
}
\value{
An \code{flFit} object that contains all fluorescence fitting results, compatible with
various plotting functions of the QurvE package.
\item{raw.x}{Raw x matrix passed to the function as \code{time} (for x_type = 'time') or \code{density} (for x_type = 'density').}
\item{raw.fl}{Raw density dataframe passed to the function as \code{data}.}
\item{flTable}{Table with fluorescence parameters and related statistics for each fluorescence curve evaluation performed by the function. This table, which is also returned by the generic \code{summary.flFit} method applied to a \code{flFit} object, is used as an input for \code{\link{fl.drFit}}.}
\item{flFittedLinear}{List of all \code{flFitLinear} objects, generated by the call of \code{\link{flFitLinear}}. Note: access to each object in the list via double brace: flFittedLinear[[#n]].}
\item{flFittedSplines}{List of all \code{flFitSpline} objects, generated by the call of \code{\link{flFitSpline}}. Note: access to each object via double brace: flFittedSplines[[#n]].}
\item{flBootSplines}{List of all \code{flBootSpline} objects, generated by the call of \code{\link{flBootSpline}}. Note: access to each object via double brace: flFittedSplines[[#n]].}
\item{control}{Object of class \code{fl.control} containing list of options passed to the function as \code{control}.}

An \code{flFit} object that contains all fluorescence fitting results, compatible with
various plotting functions of the QurvE package.
\item{raw.x}{Raw x dataframe passed to the function as \code{time} or \code{density}. If both arguments are used, \code{x} will be defined by 'x_type' in the \code{\link{fl.control}} object.}
\item{raw.fl}{Raw fluorescence dataframe passed to the function as \code{fl_data}.}
\item{flTable}{Table with fitted parameters and related statistics for each curve evaluation performed by the function. This table, which is also returned by the generic \code{summary.flFit} method applied to a \code{flFit} object, is used as an input for \code{\link{growth.drFit}} or \code{\link{fl.drFit}}.}
\item{flFittedLinear}{List of all \code{flFitLinear} objects, generated by the call of \code{\link{flFitLinear}}. Note: access to each object in the list via double brace: flFittedLinear[[#n]].}
\item{flFittedSplines}{List of all \code{flFitSpline} objects, generated by the call of \code{\link{flFitSpline}}. Note: access to each object via double brace: flFittedSplines[[#n]].}
\item{flBootSplines}{List of all \code{flBootSpline} objects, generated by the call of \code{\link{flBootSpline}}. Note: access to each object via double brace: flFittedSplines[[#n]].}
\item{control}{Object of class \code{fl.control} containing list of options passed to the function as \code{control}.}
}
\description{
\code{flFit} performs all computational fluorescence fitting operations based on the user input.
}
\seealso{
Other workflows: 
\code{\link{growth.gcFit}()},
\code{\link{growth.workflow}()}

Other fluorescence fitting functions: 
\code{\link{flBootSpline}()},
\code{\link{flFitSpline}()}

Other dose-response analysis functions: 
\code{\link{growth.drBootSpline}()},
\code{\link{growth.drFitSpline}()},
\code{\link{growth.gcFit}()},
\code{\link{growth.workflow}()}
}
\concept{dose-response analysis functions}
\concept{fluorescence fitting functions}
\concept{workflows}
