% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{grofit.param}
\alias{grofit.param}
\title{Internal function called within \code{\link{growth.gcFitModel}}.}
\usage{
grofit.param(time, data, gcID = "undefined", control)
}
\arguments{
\item{time}{Vector of the independent variable (usually time).}

\item{data}{Vector of dependent variable (usually density values).}

\item{gcID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{grofit.control} object created with \code{\link{growth.control}}, defining relevant fitting options.}
}
\description{
Internal function called within \code{\link{growth.gcFitModel}}.
}
\seealso{
Other fitting functions: 
\code{\link{growth.drBootSpline}()},
\code{\link{growth.drFitSpline}()},
\code{\link{growth.drFit}()},
\code{\link{growth.gcBootSpline}()},
\code{\link{growth.gcFitLinear}()},
\code{\link{growth.gcFitModel}()},
\code{\link{growth.gcFitSpline}()}
}
\concept{fitting functions}
