% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcBootSpline}
\alias{growth.gcBootSpline}
\title{Perform a bootstrap on the dataset followed by spline fits for each resample}
\usage{
growth.gcBootSpline(time, data, gcID = "undefined", control = growth.control())
}
\arguments{
\item{time}{Vector of the independent variable (usually: time).}

\item{data}{Vector of dependent variable (usually: density values).}

\item{gcID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{grofit.control} object created with \code{\link{growth.control}}, defining relevant fitting options.}
}
\value{
A \code{gcBootSpline} object containing a distribution of growth parameters and
a \code{gcFitSpline} object for each bootstrap sample. Use \code{\link{plot.gcBootSpline}}
to visualize all bootstrapping splines as well as the distribution of
physiological parameters.
\item{raw.time}{Raw time values provided to the function as \code{time}.}
\item{raw.data}{Raw density data for provided to the function as \code{data}.}
\item{gcID}{(Character) Identifies the tested sample.}
\item{boot.time}{Table of time values per column, resulting from each spline fit of the bootstrap.}
\item{boot.data}{Table of density values per column, resulting from each spline fit of the bootstrap.}
\item{lambda}{Vector of estimated lambda values from each bootstrap entry.}
\item{mu}{Vector of estimated mu values from each bootstrap entry.}
\item{integral}{Vector of estimated integral values from each bootstrap entry.}
\item{bootFlag}{(Logical) Indicates the success of the bootstrapping operation.}
\item{control}{Object of class \code{grofit.control} containing list of options passed to the function as \code{control}.}
}
\description{
\code{growth.gcBootSpline} resamples the density values in a dataset with replacement and performs a spline fit for each bootstrap sample.
}
\seealso{
Other growth fitting functions: 
\code{\link{grofit.param}()},
\code{\link{growth.drFit}()},
\code{\link{growth.gcFitLinear}()},
\code{\link{growth.gcFitModel}()},
\code{\link{growth.gcFitSpline}()},
\code{\link{growth.gcFit}()},
\code{\link{growth.workflow}()}
}
\concept{growth fitting functions}
