% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{growth.gcFitModel}
\alias{growth.gcFitModel}
\title{Fit different growth models to density data}
\usage{
growth.gcFitModel(time, data, gcID = "undefined", control = growth.control())
}
\arguments{
\item{gcID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{grofit.control} object created with \code{growth.control()}, defining relevant fitting options.}
}
\description{
Fit different growth models to density data
}
