% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_fits.R
\name{growth.param}
\alias{growth.param}
\title{Internal function called within \code{\link{growth.gcFitModel}}.}
\usage{
growth.param(time, data, gcID = "undefined", control)
}
\arguments{
\item{time}{Vector of the independent variable (usually time).}

\item{data}{Vector of dependent variable (usually density values).}

\item{gcID}{(Character) The name of the analyzed sample.}

\item{control}{A \code{grofit.control} object created with \code{\link{growth.control}}, defining relevant fitting options.}
}
\description{
Internal function called within \code{\link{growth.gcFitModel}}.
}
