% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_computation.R
\name{parse_data}
\alias{parse_data}
\title{Parse raw plate reader data and convert it to a format compatible with QurvE}
\usage{
parse_data(
  data.file = NULL,
  map.file = NULL,
  software = c("Gen5", "Gen6", "Chi.Bio", "GrowthProfiler", "Tecan"),
  convert.time = NULL,
  sheet.data = 1,
  sheet.map = 1,
  csvsep.data = ";",
  dec.data = ".",
  csvsep.map = ";",
  dec.map = ".",
  subtract.blank = T,
  calibration = NULL
)
}
\arguments{
\item{data.file}{(Character) A table file with extension '.xlsx', '.xls', '.csv', '.tsv', or '.txt' containing plate reader data.}

\item{map.file}{(Character) A table file in column format with 'well', 'ID', 'replicate', and 'concentration' in the first row. Used to assign sample information to wells in a plate.}

\item{software}{(Character) The name of the software/device used to export the plate reader data.}

\item{convert.time}{(\code{NULL} or string) Convert time values with a formula provided in the form \code{'y = function(x)'}.
For example: \code{convert.time = 'y = 24 * x'}}

\item{sheet.data}{(Numeric or Character) Number or name of a sheet in XLS or XLSX files containing experimental data (\emph{optional}).}

\item{sheet.map}{(Numeric or Character) Number or name of a sheet in XLS or XLSX files containing experimental data (\emph{optional}).}

\item{csvsep.data}{(Character) separator used in CSV data file (ignored for other file types).  Default: \code{";"}}

\item{dec.data}{(Character) decimal separator used in CSV, TSV or TXT data file.}

\item{csvsep.map}{(Character) separator used in CSV mapping file (ignored for other file types).  Default: \code{";"}}

\item{dec.map}{(Character) decimal separator used in CSV, TSV or TXT mapping file.}

\item{subtract.blank}{(Logical) Shall blank values be subtracted from values within the same experiment (\link{TRUE}, the default) or not (\link{FALSE}).}

\item{calibration}{(Character or \code{NULL}) Provide an equation in the form 'y = function(x)' (for example: 'y = x^2 * 0.3 - 0.5') to convert density and fluorescence values. This can be used to, e.g., convert plate reader absorbance values into \ifelse{html}{\out{OD<sub>600</sub>}}{\eqn{OD_{600}}}.
Caution!: When utilizing calibration, carefully consider whether or not blanks were subtracted to determine the calibration before selecting the input \code{subtract.blank = TRUE}.}
}
\value{
A \code{grodata} object suitable to run \code{\link{growth.workflow}}. See \code{\link{read_data}} for object structure.
}
\description{
\code{parse_data} takes a raw export file from a plate reader experiment, extracts relevant information and parses it into the format required to run \code{\link{growth.workflow}}.
}
