% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.drBootSpline}
\alias{plot.drBootSpline}
\title{Generic plot function for \code{gcBootSpline} objects.}
\usage{
\method{plot}{drBootSpline}(
  x,
  pch = 19,
  colData = 1,
  colSpline = scales::alpha("black", 0.15),
  cex.point = 1,
  cex.lab = 1.5,
  cex.axis = 1.3,
  lwd = 2,
  plot = TRUE,
  export = FALSE,
  height = 7,
  width = 9,
  out.dir = NULL,
  shiny = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{drBootSpline} object created with \code{\link{growth.drBootSpline}} or stored within a \code{grofit} or \code{drFit} object created with \code{\link{growth.workflow}} or \code{\link{growth.drFit}}, respectively.}

\item{pch}{(Numeric) Shape of the raw data symbols.}

\item{colData}{(Numeric or Character) Color used to plot the raw data.}

\item{colSpline}{(Numeric or Character) Color used to plot the splines.}

\item{cex.point}{(Numeric) Size of the raw data points.}

\item{cex.lab}{(Numeric) Font size of axis titles.}

\item{cex.axis}{(Numeric) Font size of axis annotations.}

\item{lwd}{(Numeric) Spline line width.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}).}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{shiny}{(Logical) Indicate if plot is generated within the shiny app.}

\item{...}{Further arguments to refine the generated base R plot.}
}
\description{
Generic plot function for \code{gcBootSpline} objects.
}
