% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_plots.R, R/growth_plots.R
\name{plot.drFitModel}
\alias{plot.drFitModel}
\title{Generic plot function for \code{drFitModel} objects.}
\usage{
\method{plot}{drFitModel}(
  drFitModel,
  type = c("average", "all", "bars", "none", "obs", "confidence"),
  ec50line = TRUE,
  add = FALSE,
  broken = TRUE,
  bp,
  gridsize = 200,
  log = "x",
  n.xbreaks,
  n.ybreaks,
  x.lim,
  y.lim,
  cex.point,
  cex.axis = 1,
  cex.lab = 1.5,
  col = 1,
  lwd = 2,
  lty = 2,
  xlab,
  ylab,
  legend = T,
  legendText,
  cex.legend = 1.2
)

\method{plot}{drFitModel}(
  drFitModel,
  type = c("average", "all", "bars", "none", "obs", "confidence"),
  ec50line = TRUE,
  add = FALSE,
  broken = TRUE,
  bp,
  gridsize = 200,
  log = "x",
  n.xbreaks,
  n.ybreaks,
  x.lim,
  y.lim,
  cex.point,
  cex.axis = 1,
  cex.lab = 1.5,
  col = 1,
  lwd = 2,
  lty = 2,
  xlab,
  ylab,
  legend = T,
  legendText,
  cex.legend = 1.2
)
}
\arguments{
\item{drFitModel}{object of class \code{drFitModel}, created with \code{\link{growth.drFitModel}}.}

\item{type}{(Character) Specify how to plot the data. There are currently 5 options: "average" (averages and fitted curve(s); default), "none" (only the fitted curve(s)), "obs" (only the data points), "all" (all data points and fitted curve(s)), "bars" (averages and fitted curve(s) with model-based standard errors (see Details)), and "confidence" (confidence bands for fitted curve(s)).}

\item{ec50line}{(Logical) Show pointed horizontal and vertical lines at the EC50 values (\code{TRUE}) or not (\code{FALSE}).}

\item{add}{(Logical) If \code{TRUE} then add to already existing plot.}

\item{broken}{(Logical) If TRUE the x axis is broken provided this axis is logarithmic (using functionality in the CRAN package 'plotrix').}

\item{bp}{(Numeric) Specifying the break point below which the dose is zero (the amount of stretching on the dose axis above zero in order to create the visual illusion of a logarithmic scale including 0). The default is the base-10 value corresponding to the rounded value of the minimum of the log10 values of all positive dose values. This argument is only working for logarithmic dose axes.}

\item{gridsize}{(Numeric) Number of points in the grid used for plotting the fitted curves.}

\item{log}{(Character) String which contains '"x"' if the x axis is to be logarithmic, '"y"' if the y axis is to be logarithmic and '"xy"' or '"yx"' if both axes are to be logarithmic. The default is "x". The empty string "" yields the original axes.}

\item{n.xbreaks}{(Numeric) Number of breaks on the x-axis (if not log-transformed). The breaks are generated using \code{pretty}. Thus, the final number of breaks can deviate from the user input.}

\item{n.ybreaks}{(Numeric) Number of breaks on the y-axis (if not log-transformed). The breaks are generated using \code{pretty}. Thus, the final number of breaks can deviate from the user input.}

\item{x.lim}{(Numeric vector with two elements) Optional: Provide the lower (\code{l}) and upper (\code{u}) bounds on the x-axis of both growth curve and derivative plots as a vector in the form \code{c(l, u)}. If only the lower or upper bound should be fixed, provide \code{c(l, NA)} or \code{c(NA, u)}, respectively.}

\item{y.lim}{(Numeric vector with two elements) Optional: Provide the lower (\code{l}) and upper (\code{u}) bounds on y-axis of the growth curve plot as a vector in the form \code{c(l, u)}. If only the lower or upper bound should be fixed, provide \code{c(l, NA)} or \code{c(NA, u)}, respectively.}

\item{cex.point}{(Numeric) Size of the raw data points.}

\item{cex.axis}{(Numeric) Font size of axis annotations.}

\item{cex.lab}{(Numeric) Font size of axis titles.}

\item{col}{(Logical or a vector of colors) If \code{TRUE} default colours are used. If \code{FALSE} (default) no colors are used.}

\item{lwd}{(Numeric) Line width.}

\item{lty}{(Numeric) Specify the line type.}

\item{xlab}{(Character) An optional label for the x axis.}

\item{ylab}{(Character) An optional label for the y axis.}

\item{legend}{(Logical) If \code{TRUE} a legend is displayed.}

\item{legendText}{a character string or vector of character strings specifying the legend text (the position of the upper right corner of the legend box).}

\item{cex.legend}{numeric specifying the legend text size.}

\item{drFittedModel}{Object of class \code{drFitModel}, created with \code{\link{fl.drFitModel}}.}

\item{pch}{(Numeric) Size of the raw data circles.}

\item{colData}{(Numeric or Character) Color used to plot the raw data.}

\item{colSpline}{(Numeric or Character) Color used to plot the splines.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}).}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{...}{Further arguments to refine the generated base R plot.}
}
\description{
Generic plot function for \code{drFitModel} objects.

Generic plot function for \code{drFitModel} objects.
}
