% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.gcFitModel}
\alias{plot.gcFitModel}
\title{Generic plot function for \code{gcBootSpline} objects.}
\usage{
\method{plot}{gcFitModel}(
  gcFittedModel,
  raw = TRUE,
  pch = 1,
  colData = 1,
  equation = TRUE,
  eq.size = 1,
  colModel = ggplot2::alpha("forestgreen", 0.85),
  basesize = 16,
  cex.point = 2,
  lwd = 0.7,
  n.ybreaks = NULL,
  plot = TRUE,
  export = FALSE,
  height = 8,
  width = 6,
  out.dir = NULL,
  ...
)
}
\arguments{
\item{gcFittedModel}{A \code{gcFittedModel} object created with \code{\link{growth.gcFitModel}} or stored within a \code{grofit} or \code{gcFit} object created with \code{\link{growth.workflow}} or \code{\link{growth.gcFit}}, respectively.}

\item{raw}{(Logical) Show the raw data within the plot (\code{TRUE}) or not (\code{FALSE}).}

\item{pch}{(Numeric) Symbol used to plot data points.}

\item{colData}{(Numeric or Character) Color used to plot the raw data.}

\item{equation}{(Logical) Show the equation of the fitted model within the plot (\code{TRUE}) or not (\code{FALSE}).}

\item{colModel}{(Numeric or Character) Color used to plot the fitted model.}

\item{basesize}{(Numeric) Base font size.}

\item{cex.point}{(Numeric) Size of the raw data points.}

\item{lwd}{(Numeric) Spline line width.}

\item{n.ybreaks}{(Numeric) Number of breaks on the y-axis. The breaks are generated using \code{scales::pretty_breaks}. Thus, the final number of breaks can deviate from the user input.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}). If \code{FALSE}, a ggplot object is returned.}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{...}{Further arguments to refine the generated \code{ggplot2} plot.}
}
\description{
Plot the results of a parametric model fit on density vs. time data
}
