% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_plots.R
\name{plot.gcFitSpline}
\alias{plot.gcFitSpline}
\title{Generic plot function for \code{gcFitSpline} objects.}
\usage{
\method{plot}{gcFitSpline}(
  gcFittedSpline,
  add = FALSE,
  raw = TRUE,
  slope = TRUE,
  deriv = T,
  spline = T,
  log.y = T,
  pch = 1,
  colData = 1,
  colSpline = "dodgerblue3",
  basesize = 16,
  cex.point = 2,
  lwd = 0.7,
  y.lim = NULL,
  x.lim = NULL,
  y.lim.deriv = NULL,
  n.ybreaks = 6,
  plot = TRUE,
  export = FALSE,
  width = 8,
  height = ifelse(deriv == TRUE, 8, 6),
  out.dir = NULL,
  ...
)
}
\arguments{
\item{add}{(Logical) Shall the fitted spline be added to an existing plot? \code{TRUE} is used internally by \code{\link{plot.gcBootSpline}}.}

\item{raw}{(Logical) Display raw density as circles (\code{TRUE}) or not (\code{FALSE}).}

\item{slope}{(Logical) Show the slope at the maximum growth rate (\code{TRUE}) or not (\code{FALSE}).}

\item{deriv}{(Logical) Show the derivative (i.e., slope) over time in a secondary plot (\code{TRUE}) or not (\code{FALSE}).}

\item{spline}{(Logical) Only for \code{add = TRUE}: add the current spline to the existing plot (\code{FALSE}).}

\item{log.y}{(Logical) Log-transform the y-axis (\code{TRUE}) or not (\code{FALSE}).}

\item{pch}{(Numeric) Symbol used to plot data points.}

\item{colData}{(Numeric or character) Contour color of the raw data circles.}

\item{colSpline}{(Numeric or character) Spline line colour.}

\item{basesize}{(Numeric) Base font size.}

\item{cex.point}{(Numeric) Size of the raw data points.}

\item{lwd}{(Numeric) Spline line width.}

\item{y.lim}{(Numeric vector with two elements) Optional: Provide the lower (\code{l}) and upper (\code{u}) bounds on y-axis of the growth curve plot as a vector in the form \code{c(l, u)}. If only the lower or upper bound should be fixed, provide \code{c(l, NA)} or \code{c(NA, u)}, respectively.}

\item{x.lim}{(Numeric vector with two elements) Optional: Provide the lower (\code{l}) and upper (\code{u}) bounds on the x-axis of both growth curve and derivative plots as a vector in the form \code{c(l, u)}. If only the lower or upper bound should be fixed, provide \code{c(l, NA)} or \code{c(NA, u)}, respectively.}

\item{y.lim.deriv}{(Numeric vector with two elements) Optional: Provide the lower (\code{l}) and upper (\code{u}) bounds on the y-axis of the derivative plot as a vector in the form \code{c(l, u)}. If only the lower or upper bound should be fixed, provide \code{c(l, NA)} or \code{c(NA, u)}, respectively.}

\item{n.ybreaks}{(Numeric) Number of breaks on the y-axis. The breaks are generated using \code{scales::pretty_breaks}. Thus, the final number of breaks can deviate from the user input.}

\item{plot}{(Logical) Show the generated plot in the \code{Plots} pane (\code{TRUE}) or not (\code{FALSE}). If \code{FALSE}, a ggplot object is returned.}

\item{export}{(Logical) Export the generated plot as PDF and PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{width}{(Numeric) Width of the exported image in inches.}

\item{height}{(Numeric) Height of the exported image in inches.}

\item{out.dir}{(Character) Name or path to a folder in which the exported files are stored. If \code{NULL}, a "Plots" folder is created in the current working directory to store the files in.}

\item{...}{}

\item{gcFitSpline}{object of class \code{gcFitSpline}, created with \code{\link{growth.gcFitSpline}}.}
}
\description{
code{plot.gcFitSpline} generates the spline fit plot for a single sample.
}
