% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R
\name{summary.flBootSpline}
\alias{summary.flBootSpline}
\title{Generic summary function for flBootSpline objects}
\usage{
\method{summary}{flBootSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{flBootSpline}}

\item{...}{}
}
\value{
A dataframe with statistical parameters extracted from the dose-response bootstrapping analysis.
}
\description{
Generic summary function for flBootSpline objects
}
