% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluorescence_summaries.R
\name{summary.flFitSpline}
\alias{summary.flFitSpline}
\title{Generic summary function for flFitSpline objects}
\usage{
\method{summary}{flFitSpline}(object, ...)
}
\arguments{
\item{object}{object of class \code{flFitSpline}}

\item{...}{Additional arguments. This has currently no effect and is only meant to fulfill the requirements of a generic function.}
}
\value{
A dataframe with parameters extracted from the nonparametric fit.
}
\description{
Generic summary function for flFitSpline objects
}
