% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{xgx_minor_breaks_log10}
\alias{xgx_minor_breaks_log10}
\title{Sets the default minor_breaks for log10 scales (from R package 'xgxr')}
\usage{
xgx_minor_breaks_log10(data_range)
}
\arguments{
\item{data_range}{range of the data}
}
\value{
numeric vector of breaks
}
\description{
\code{xgx_minor_breaks_log10} sets nice minor_breaks for log10 scale.
}
\examples{
xgx_minor_breaks_log10(c(1, 1000))
xgx_minor_breaks_log10(c(0.001, 100))
xgx_minor_breaks_log10(c(1e-4, 1e4))
xgx_minor_breaks_log10(c(1e-9, 1e9))
xgx_minor_breaks_log10(c(1, 2))
xgx_minor_breaks_log10(c(1, 5))
xgx_minor_breaks_log10(c(1, 10))
xgx_minor_breaks_log10(c(1, 100))
xgx_minor_breaks_log10(c(1, 1.01))
xgx_minor_breaks_log10(c(1, 1.0001))
print(xgx_minor_breaks_log10(c(1, 1.000001)), digits = 10)

}
